refactor: #102 프로모션 상세 조회 API '저장 수', '내 저장 여부' 조회 로직 추가

+++ src/main/kotlin/com/damaba/damaba/domain/promotion/PromotionDetail.kt
@@ -19,7 +19,7 @@ data class PromotionDetail(
     val startedAt: LocalDate?,
     val endedAt: LocalDate?,
     val viewCount: Long,
-    val saveCount: Int,
+    val saveCount: Long,
     val isSaved: Boolean,
     val photographyTypes: Set<PhotographyType>,
     val images: List<Image>,
